Fix clang-tidy warnings

+++ torch/csrc/distributed/rpc/tensorpipe_agent.cpp
@@ -624,13 +624,13 @@ void TensorPipeAgent::sendCompletedResponseMessage(
   if (!futureResponseMessage.hasError()) {
     c10::intrusive_ptr<Message> responseMessage =
         futureResponseMessage.value().toCustomClass<Message>();
-    responseMessage->setId(messageId);
+    responseMessage->setId(static_cast<int64_t>(messageId));
 
     std::vector<c10::Device> devices;
     try {
       devices = getDevicesForRemote(pipe->getRemoteName(), *responseMessage);
     } catch (const std::exception& e) {
-      responseMessage = createExceptionResponse(e.what(), messageId);
+      responseMessage = createExceptionResponse(e.what(), static_cast<int64_t>(messageId));
     }
 
     for (const auto& tensor : responseMessage->tensors()) {
@@ -652,7 +652,7 @@ void TensorPipeAgent::sendCompletedResponseMessage(
                   oss.str(),
                   "which is not yet supported. Please file a feature request "
                   "issue in PyTorch GitHub repo."),
-              messageId);
+              static_cast<int64_t>(messageId));
           break;
         }
       }
@@ -681,7 +681,7 @@ void TensorPipeAgent::sendCompletedResponseMessage(
     pipeWrite(
         pipe,
         createExceptionResponse(
-            futureResponseMessage.tryRetrieveErrorMessage(), messageId),
+            futureResponseMessage.tryRetrieveErrorMessage(), static_cast<int64_t>(messageId)),
         /* devices */ {},
         std::move(streams),
         [this, pipe, messageId](const tensorpipe::Error& error) {
